move submission utils to new file
